+++ modules/main/src/main/java/com/nisovin/shopkeepers/lang/Messages.java
@@ -155,11 +155,6 @@ public class Messages extends Config {
 			"Changes the coat pattern",
 			"of the horse"
 	));
-	public static String buttonHorseSaddle = c("&aToggle horse saddle");
-	public static List<String> buttonHorseSaddleLore = c(Arrays.asList(
-			"Toggles the saddle",
-			"of the horse"
-	));
 	public static String buttonHorseArmor = c("&aChoose horse armor");
 	public static List<String> buttonHorseArmorLore = c(Arrays.asList(
 			"Changes the armor",
@@ -209,10 +204,6 @@ public class Messages extends Config {
 	public static String buttonParrotVariant = c("&aChoose parrot variant");
 	public static List<String> buttonParrotVariantLore = c(Arrays.asList(
 			"Changes the parrot's look"
-	));
-	public static String buttonPigSaddle = c("&aToggle pig saddle");
-	public static List<String> buttonPigSaddleLore = c(Arrays.asList(
-			"Toggles the pig's saddle"
 	));
 	public static String buttonPigVariant = c("&aChoose pig variant");
 	public static List<String> buttonPigVariantLore = c(Arrays.asList(
@@ -331,6 +322,10 @@ public class Messages extends Config {
 	public static List<String> buttonWolfVariantLore = c(Arrays.asList(
 			"Changes the wolf's look"
 	));
+	public static String buttonSaddle = c("&aToggle saddle");
+	public static List<String> buttonSaddleLore = c(Arrays.asList(
+			"Toggles the mob's saddle"
+	));
 
 	public static String equipmentEditorTitle = c("Equipment Editor");
 	public static List<String> equipmentSlotLore = c(Arrays.asList(
@@ -346,6 +341,7 @@ public class Messages extends Config {
 	public static String equipmentSlotChest = c("&aChest");
 	public static String equipmentSlotHead = c("&aHead");
 	public static String equipmentSlotBody = c("&aBody");
+	public static String equipmentSlotSaddle = c("&aSaddle");
 
 	public static String adminSignShopLine1 = c("&2[SHOP]");
 	public static String adminSignShopLine2 = c("&7{shopName}");