+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopobjects/living/types/HorseShop.java
@@ -30,12 +30,11 @@ import com.nisovin.shopkeepers.util.data.property.BasicProperty;
 import com.nisovin.shopkeepers.util.data.property.Property;
 import com.nisovin.shopkeepers.util.data.property.value.PropertyValue;
 import com.nisovin.shopkeepers.util.data.serialization.InvalidDataException;
-import com.nisovin.shopkeepers.util.data.serialization.java.BooleanSerializers;
 import com.nisovin.shopkeepers.util.data.serialization.java.EnumSerializers;
 import com.nisovin.shopkeepers.util.inventory.ItemUtils;
 import com.nisovin.shopkeepers.util.java.EnumUtils;
 
-public class HorseShop extends BabyableShop<Horse> {
+public class HorseShop extends AbstractHorseShop<Horse> {
 
 	public enum HorseArmor {
 
@@ -66,11 +65,6 @@ public class HorseShop extends BabyableShop<Horse> {
 			.defaultValue(Horse.Style.NONE)
 			.build();
 
-	public static final Property<Boolean> SADDLE = new BasicProperty<Boolean>()
-			.dataKeyAccessor("saddle", BooleanSerializers.LENIENT)
-			.defaultValue(false)
-			.build();
-
 	public static final Property<@Nullable HorseArmor> ARMOR = new BasicProperty<@Nullable HorseArmor>()
 			.dataKeyAccessor("armor", EnumSerializers.lenient(HorseArmor.class))
 			.nullable() // Null indicates 'no armor'
@@ -82,9 +76,6 @@ public class HorseShop extends BabyableShop<Horse> {
 			.build(properties);
 	private final PropertyValue<Horse.Style> styleProperty = new PropertyValue<>(STYLE)
 			.onValueChanged(Unsafe.initialized(this)::applyStyle)
-			.build(properties);
-	private final PropertyValue<Boolean> saddleProperty = new PropertyValue<>(SADDLE)
-			.onValueChanged(Unsafe.initialized(this)::applySaddle)
 			.build(properties);
 	private final PropertyValue<@Nullable HorseArmor> armorProperty = new PropertyValue<>(ARMOR)
 			.onValueChanged(Unsafe.initialized(this)::applyArmor)
@@ -104,7 +95,6 @@ public class HorseShop extends BabyableShop<Horse> {
 		super.load(shopObjectData);
 		colorProperty.load(shopObjectData);
 		styleProperty.load(shopObjectData);
-		saddleProperty.load(shopObjectData);
 		armorProperty.load(shopObjectData);
 	}
 
@@ -113,7 +103,6 @@ public class HorseShop extends BabyableShop<Horse> {
 		super.save(shopObjectData, saveAll);
 		colorProperty.save(shopObjectData);
 		styleProperty.save(shopObjectData);
-		saddleProperty.save(shopObjectData);
 		armorProperty.save(shopObjectData);
 	}
 
@@ -122,7 +111,6 @@ public class HorseShop extends BabyableShop<Horse> {
 		super.onSpawn();
 		this.applyColor();
 		this.applyStyle();
-		this.applySaddle();
 		this.applyArmor();
 	}
 
@@ -131,7 +119,6 @@ public class HorseShop extends BabyableShop<Horse> {
 		List<Button> editorButtons = super.createEditorButtons();
 		editorButtons.add(this.getColorEditorButton());
 		editorButtons.add(this.getStyleEditorButton());
-		editorButtons.add(this.getSaddleEditorButton());
 		editorButtons.add(this.getArmorEditorButton());
 		return editorButtons;
 	}
@@ -268,54 +255,6 @@ public class HorseShop extends BabyableShop<Horse> {
 			) {
 				boolean backwards = clickEvent.isRightClick();
 				cycleStyle(backwards);
-				return true;
-			}
-		};
-	}
-
-	// SADDLE
-
-	public boolean hasSaddle() {
-		return saddleProperty.getValue();
-	}
-
-	public void setSaddle(boolean saddle) {
-		saddleProperty.setValue(saddle);
-	}
-
-	public void cycleSaddle() {
-		this.setSaddle(!this.hasSaddle());
-	}
-
-	private void applySaddle() {
-		Horse entity = this.getEntity();
-		if (entity == null) return; // Not spawned
-
-		entity.getInventory().setSaddle(this.hasSaddle() ? new ItemStack(Material.SADDLE) : null);
-	}
-
-	private ItemStack getSaddleEditorItem() {
-		ItemStack iconItem = new ItemStack(Material.SADDLE);
-		ItemUtils.setDisplayNameAndLore(iconItem,
-				Messages.buttonHorseSaddle,
-				Messages.buttonHorseSaddleLore
-		);
-		return iconItem;
-	}
-
-	private Button getSaddleEditorButton() {
-		return new ShopkeeperActionButton() {
-			@Override
-			public @Nullable ItemStack getIcon(EditorSession editorSession) {
-				return getSaddleEditorItem();
-			}
-
-			@Override
-			protected boolean runAction(
-					EditorSession editorSession,
-					InventoryClickEvent clickEvent
-			) {
-				cycleSaddle();
 				return true;
 			}
 		};