+++ modules/main/src/main/java/com/nisovin/shopkeepers/util/inventory/InventoryUtils.java
@@ -7,7 +7,6 @@ import org.bukkit.Bukkit;
 import org.bukkit.entity.HumanEntity;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.InventoryType;
-import org.bukkit.inventory.EquipmentSlot;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.InventoryView;
 import org.bukkit.inventory.ItemStack;
@@ -666,31 +665,6 @@ public final class InventoryUtils {
 		Bukkit.getScheduler().runTask(ShopkeepersPlugin.getInstance(), () -> {
 			inventory.setItem(slot, itemStack); // This copies the item internally
 		});
-	}
-
-	// TODO Replace this with the corresponding Bukkit API method added in late 1.15.2. See
-	// https://hub.spigotmc.org/stash/projects/SPIGOT/repos/bukkit/commits/da9ef3c55fa3bce91f7fdcd77d50171be7297d7d
-	// Note: MC 1.20.5 added EquipmentSlot.Body. E.g. used for horse armor. Not relevant for
-	// PlayerInventory.
-	public static @Nullable ItemStack getItem(PlayerInventory playerInventory, EquipmentSlot slot) {
-		Validate.notNull(playerInventory, "playerInventory is null");
-		Validate.notNull(slot, "slot is null");
-		switch (slot) {
-		case HAND:
-			return playerInventory.getItemInMainHand();
-		case OFF_HAND:
-			return playerInventory.getItemInOffHand();
-		case FEET:
-			return playerInventory.getBoots();
-		case LEGS:
-			return playerInventory.getLeggings();
-		case CHEST:
-			return playerInventory.getChestplate();
-		case HEAD:
-			return playerInventory.getHelmet();
-		default:
-			return null;
-		}
 	}
 
 	private InventoryUtils() {