+++ modules/main/src/main/java/com/nisovin/shopkeepers/villagers/BlockZombieVillagerCuringListener.java
@@ -14,7 +14,6 @@ import org.bukkit.inventory.ItemStack;
 
 import com.nisovin.shopkeepers.lang.Messages;
 import com.nisovin.shopkeepers.util.bukkit.TextUtils;
-import com.nisovin.shopkeepers.util.inventory.InventoryUtils;
 import com.nisovin.shopkeepers.util.logging.Log;
 
 /**
@@ -31,7 +30,7 @@ public class BlockZombieVillagerCuringListener implements Listener {
 	void onZombieVillagerCureStarted(PlayerInteractEntityEvent event) {
 		if (!(event.getRightClicked() instanceof ZombieVillager)) return;
 		Player player = event.getPlayer();
-		ItemStack itemInHand = InventoryUtils.getItem(player.getInventory(), event.getHand());
+		ItemStack itemInHand = player.getInventory().getItem(event.getHand());
 		if (itemInHand != null && itemInHand.getType() == Material.GOLDEN_APPLE) {
 			// Prevent curing:
 			Log.debug(() -> "Preventing zombie villager curing at "